move validation check, reenable passing strs for enums
