+++ fides/minimize.py
@@ -18,12 +18,11 @@ from .hessian_approximation import (
     HybridFraction, FX, IterativeHessianApproximation, TSSM, GNSBFGS
 )
 from .constants import (
-    Options, ExitFlag, DEFAULT_OPTIONS, StepBackStrategy, SubSpaceDim
+    Options, ExitFlag, DEFAULT_OPTIONS, StepBackStrategy, SubSpaceDim,
+    validate_options
 )
 from .logging import create_logger
 from collections import defaultdict
-from numbers import Real, Integral
-from pathlib import PosixPath, WindowsPath
 from typing import Callable, Dict, Optional, Tuple, Union, List
 
 
@@ -844,38 +843,3 @@ class Optimizer:
 def _min_max_evs(mat: np.ndarray):
     evs = np.linalg.eigvals(mat)
     return np.real(np.min(evs)), np.real(np.max(evs))
-
-
-def validate_options(options: Dict):
-    """Check if the chosen options are valid"""
-    expected_types = {
-        Options.MAXITER: Integral,
-        Options.MAXTIME: Real,
-        Options.FATOL: Real,
-        Options.FRTOL: Real,
-        Options.XTOL: Real,
-        Options.GATOL: Real,
-        Options.GRTOL: Real,
-        Options.SUBSPACE_DIM: SubSpaceDim,
-        Options.STEPBACK_STRAT: StepBackStrategy,
-        Options.THETA_MAX: Real,
-        Options.DELTA_INIT: Real,
-        Options.MU: Real,
-        Options.ETA: Real,
-        Options.GAMMA1: Real,
-        Options.GAMMA2: Real,
-        Options.HISTORY_FILE: (str, PosixPath, WindowsPath),
-    }
-    for option_key, option_value in options.items():
-        try:
-            option = Options(option_key)
-        except ValueError:
-            raise ValueError(f'{option_key} is not a valid options field.')
-
-        expected_type = expected_types[option]
-        if not isinstance(option_value, expected_type):
-            if expected_type == Integral and int(option_value) == option_value:
-                continue
-            raise TypeError(f'Type mismatch for option {option_key}. '
-                            f'Expected {expected_type} but got '
-                            f'{type(option_value)}')
